+++ backend/canisters/group/api/src/queries/events_window.rs
@@ -8,7 +8,6 @@ pub struct Args {
     pub mid_point: MessageIndex,
     pub max_messages: u32,
     pub max_events: u32,
-    pub invite_code: Option<u64>,
     pub latest_client_event_index: Option<EventIndex>,
 }