+++ backend/canisters/group/api/src/updates/decline_invitation.rs
@@ -6,11 +6,7 @@ pub type Args = Empty;
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub enum Response {
-    Success(SuccessResult),
-    NotAuthorized,
-}
-
-#[derive(CandidType, Serialize, Deserialize, Debug)]
-pub struct SuccessResult {
-    pub code: Option<u64>,
+    Success,
+    NotInvited,
+    InternalError(String),
 }