+++ backend/libraries/types/src/events.rs
@@ -45,6 +45,7 @@ pub enum ChatEvent {
     ChatUnfrozen(ChatUnfrozen),
     EventsTimeToLiveUpdated(EventsTimeToLiveUpdated),
     GroupGateUpdated(GroupGateUpdated),
+    UsersInvited(UsersInvited),
 }
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
@@ -117,6 +118,8 @@ pub struct UsersUnblocked {
 #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
 pub struct ParticipantJoined {
     pub user_id: UserId,
+    #[serde(default)]
+    pub invited_by: Option<UserId>,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
@@ -255,3 +258,9 @@ pub struct GroupGateUpdated {
 
 #[derive(CandidType, Serialize, Deserialize, Copy, Clone, Debug)]
 pub struct DirectChatCreated {}
+
+#[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+pub struct UsersInvited {
+    pub user_ids: Vec<UserId>,
+    pub invited_by: UserId,
+}